requirements.txt
openai
gtts
pygame
sounddevice
vosk
python-dotenv
Pillow